+++ packages/emoji-picker/src/Context.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import { LATEST_DATASET_VERSION } from 'interweave-emoji';
+import { EmojiDataManager, LATEST_DATASET_VERSION } from 'interweave-emoji';
 import { CONTEXT_CLASSNAMES, CONTEXT_MESSAGES } from './constants';
 import { Context } from './types';
 
 export default React.createContext<Context>({
   classNames: CONTEXT_CLASSNAMES,
+  emojiData: EmojiDataManager.getInstance('en'),
   emojiLargeSize: 0,
   emojiPadding: 0,
   emojiPath: '{{hexcode}}',