+++ packages/emoji-picker/src/Group.tsx
@@ -1,7 +1,6 @@
-import React, { useContext } from 'react';
-import camelCase from 'lodash/camelCase';
-import { GROUP_KEY_COMMONLY_USED } from './constants';
+import React, { useCallback, useContext } from 'react';
 import Context from './Context';
+import useGroupMessage from './hooks/useGroupMessage';
 import { CommonMode, GroupKey } from './types';
 
 export interface GroupProps {
@@ -13,26 +12,24 @@ export interface GroupProps {
 }
 
 export default function Group({ active, children, commonMode, group, onSelect }: GroupProps) {
-  const { classNames, messages } = useContext(Context);
-  const key = camelCase(group === GROUP_KEY_COMMONLY_USED ? commonMode : group);
+  const { classNames } = useContext(Context);
   const className = [classNames.group];
+  const title = useGroupMessage(group, commonMode);
 
   if (active) {
     className.push(classNames.groupActive);
   }
 
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-    onSelect(group, event);
-  };
+  const handleClick = useCallback(
+    (event: React.MouseEvent<HTMLButtonElement>) => {
+      event.stopPropagation();
+      onSelect(group, event);
+    },
+    [group, onSelect],
+  );
 
   return (
-    <button
-      className={className.join(' ')}
-      title={messages[key]}
-      type="button"
-      onClick={handleClick}
-    >
+    <button className={className.join(' ')} title={title} type="button" onClick={handleClick}>
       {children}
     </button>
   );