+++ packages/emoji-picker/src/SkinTone.tsx
@@ -1,8 +1,9 @@
-import React, { useContext } from 'react';
+import React, { useCallback, useContext } from 'react';
 import camelCase from 'lodash/camelCase';
 import upperFirst from 'lodash/upperFirst';
 import { SKIN_COLORS } from './constants';
 import Context from './Context';
+import { useTitleFormat } from './hooks/useTitleFormat';
 import { SkinToneKey } from './types';
 
 export interface SkinToneProps {
@@ -17,22 +18,26 @@ export default function SkinTone({ active, children, skinTone, onSelect }: SkinT
   const className = [classNames.skinTone];
   const color = SKIN_COLORS[skinTone];
   const key = camelCase(skinTone);
+  const title = useTitleFormat(messages[`skin${upperFirst(key)}`]);
 
   if (active) {
     className.push(classNames.skinToneActive);
   }
 
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.stopPropagation();
-    onSelect(skinTone, event);
-  };
+  const handleClick = useCallback(
+    (event: React.MouseEvent<HTMLButtonElement>) => {
+      event.stopPropagation();
+      onSelect(skinTone, event);
+    },
+    [skinTone, onSelect],
+  );
 
   return (
     <button
       className={className.join(' ')}
       data-skin-color={color}
       data-skin-tone={skinTone}
-      title={messages[`skin${upperFirst(key)}`]}
+      title={title}
       type="button"
       onClick={handleClick}
     >