+++ packages/emoji/src/constants.ts
@@ -1,5 +1,5 @@
-// v13 isnt available on platforms yet
-export const MAX_EMOJI_VERSION = 12.1;
-
-// v5 is for emoji version v13 (above)
-export const LATEST_DATASET_VERSION = '4.2.1';
+// We hardcode these values so that newer emojis
+// are not displayed for devices that do not support
+// them yet. Support cycle is twice a year.
+export const MAX_EMOJI_VERSION = 13.1;
+export const LATEST_DATASET_VERSION = '6.1.0';