+++ packages/emoji/src/useEmojiData.tsx
@@ -1,10 +1,28 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-
-import { useEffect, useState } from 'react';
-import { Emoji, fetchFromCDN } from 'emojibase';
+import { useEffect, useMemo, useState } from 'react';
+import { fetchEmojis, fetchMetadata, Locale, ShortcodePreset } from 'emojibase';
 import { LATEST_DATASET_VERSION } from './constants';
 import EmojiDataManager from './EmojiDataManager';
 import { CanonicalEmoji, Source, UseEmojiDataOptions } from './types';
+
+export function determinePresets(locale: string, shortcodes: string[]): string[] {
+  const presets: string[] = [];
+
+  if (shortcodes.length === 0) {
+    presets.push('emojibase');
+  } else {
+    presets.push(...shortcodes);
+  }
+
+  if (
+    !locale.startsWith('en') &&
+    presets.includes('emojibase') &&
+    !presets.includes('en/emojibase')
+  ) {
+    presets.push('en/emojibase');
+  }
+
+  return presets;
+}
 
 const promises = new Map<string, Promise<CanonicalEmoji[]>>();
 
@@ -14,48 +32,37 @@ export function resetLoaded() {
   }
 }
 
-function loadEmojis(locale: string, version: string, compact: boolean): Promise<CanonicalEmoji[]> {
-  const set = compact ? 'compact' : 'data';
-  const key = `${locale}:${version}:${set}`;
-  const stubRequest =
-    (process.env.NODE_ENV === 'test' && !process.env.INTERWEAVE_ALLOW_FETCH_EMOJI) ||
-    typeof (global as any).fetch === 'undefined';
+function loadEmojis(
+  locale: Locale,
+  version: string,
+  shortcodes: (string | ShortcodePreset)[],
+  compact: boolean,
+): Promise<CanonicalEmoji[]> {
+  const key = `${locale}:${version}:${compact ? 'compact' : 'data'}`;
+  const instance = EmojiDataManager.getInstance(locale);
 
-  // Return as we've already loaded or are loading data
+  // Return as we're currently loading data
   if (promises.has(key)) {
     return promises.get(key)!;
   }
 
-  // Otherwise, start loading emoji data from the CDN
-  let request: Promise<Emoji[]>;
-
-  if (stubRequest) {
-    let testData: Emoji[];
-
-    try {
-      // We must use a variable here, otherwise webpack attempts to include it in the bundle.
-      // If that happens and the module does not exist, it will throw a warning.
-      // https://github.com/webpack/webpack/issues/8826
-      // https://github.com/webpack/webpack/issues/4175
-      const requireFunc =
-        typeof __webpack_require__ === 'function' ? __non_webpack_require__ : require;
-
-      testData = requireFunc('emojibase-test-utils/test-data.json');
-    } catch {
-      testData = [];
-    }
-
-    request = Promise.resolve(testData);
-  } else {
-    request = fetchFromCDN<Emoji>(`${locale}/${set}.json`, version);
+  // Data has been loaded elsewhere
+  if (instance.data.length > 0) {
+    return Promise.resolve(instance.getData());
   }
+
+  // Otherwise, start loading emoji data from the CDN
+  const request = Promise.all([
+    // @ts-expect-error
+    fetchEmojis(locale, { compact, flat: false, shortcodes, version }) as Emoji[],
+    fetchMetadata(locale, { version }),
+  ]);
 
   promises.set(
     key,
-    request.then((response) => {
-      const instance = EmojiDataManager.getInstance(locale);
-
-      instance.parseEmojiData(response);
+    request.then(([emojis, messages]) => {
+      instance.parseEmojiData(emojis);
+      instance.parseMessageData(messages);
 
       return instance.getData();
     }),
@@ -68,9 +75,12 @@ export default function useEmojiData({
   avoidFetch = false,
   compact = false,
   locale = 'en',
+  shortcodes = ['emojibase'],
   throwErrors = false,
   version = LATEST_DATASET_VERSION,
 }: UseEmojiDataOptions = {}): [CanonicalEmoji[], Source, EmojiDataManager] {
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const shortcodePresets = useMemo(() => determinePresets(locale, shortcodes), shortcodes);
   const [emojis, setEmojis] = useState<CanonicalEmoji[]>([]);
   const [error, setError] = useState<Error>();
 
@@ -78,7 +88,7 @@ export default function useEmojiData({
     let mounted = true;
 
     if (!avoidFetch) {
-      loadEmojis(locale, version, compact)
+      loadEmojis(locale, version, shortcodePresets, compact)
         .then((loadedEmojis) => {
           if (mounted) {
             setEmojis(loadedEmojis);
@@ -92,7 +102,7 @@ export default function useEmojiData({
     return () => {
       mounted = false;
     };
-  }, [avoidFetch, compact, locale, version]);
+  }, [avoidFetch, compact, locale, shortcodePresets, version]);
 
   if (error && throwErrors) {
     throw error;