+++ tests/setup.ts
@@ -1,12 +1,13 @@
-import { loadEmojiData } from 'emojibase-test-utils';
+import { loadEmojiData, loadMeta, loadShortcodes } from 'emojibase-test-utils';
 import { EmojiDataManager } from 'interweave-emoji';
 import EmojiDataSourceManager from '../packages/emoji/src/EmojiDataManager';
 
-const data = loadEmojiData();
+const data = loadEmojiData([loadShortcodes()]);
+const messages = loadMeta();
 
 // Bootstrap our emoji data using the official en dataset
 EmojiDataSourceManager.getInstance('en').parseEmojiData(data);
-EmojiDataManager.getInstance('en').parseEmojiData(data);
+EmojiDataSourceManager.getInstance('en').parseMessageData(messages);
 
-// @ts-ignore
-global.regeneratorRuntime = require('regenerator-runtime');
+EmojiDataManager.getInstance('en').parseEmojiData(data);
+EmojiDataManager.getInstance('en').parseMessageData(messages);